guibai

+++ 09_tong_quan_web/exercise/product_discount_calculator/src/DiscountServlet.java
@@ -9,14 +9,14 @@ import java.io.PrintWriter;
 @WebServlet(name = "DiscountServlet", urlPatterns = "/discount")
 public class DiscountServlet extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        float price =Float.parseFloat(request.getParameter("price"));
-        int discount =Integer.parseInt(request.getParameter("discount"));
-        float result = price - price * discount/100;
+        float price = Float.parseFloat(request.getParameter("price"));
+        float discount = Float.parseFloat(request.getParameter("discount"));
+        float result = price - price * discount / 100;
         PrintWriter printWriter = response.getWriter();
         printWriter.println("<html>");
-        printWriter.println("<h1>price: " + price+ "</h1>");
-        printWriter.println("<h1>discount: " + discount+ "</h1>");
-        printWriter.println("<h1>result: " + result+ "</h1>");
+        printWriter.println("<h1>price: " + price + "</h1>");
+        printWriter.println("<h1>discount: " + discount + "</h1>");
+        printWriter.println("<h1>result: " + result + "</h1>");
         printWriter.println("</html>");
     }
